+++ packages/stacked/example/lib/app/app.locator.dart
@@ -8,6 +8,7 @@
 
 import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
+import 'package:stacked_themes/stacked_themes.dart';
 
 import '../services/epoch_service.dart';
 import '../services/information_service.dart';
@@ -22,6 +23,7 @@ void setupLocator() {
   locator.registerLazySingleton(() => InformationService());
   locator.registerLazySingleton(() => NavigationService());
   locator.registerLazySingleton(() => EpochService());
+  locator.registerLazySingleton(() => ThemeService.getInstance());
   locator.registerSingleton(HistoryViewModel());
   locator.registerSingleton(FavoritesViewModel());
 }