+++ packages/stacked/example/lib/app/app.router.dart
@@ -14,6 +14,7 @@ import '../ui/bottom_nav/bottom_nav_example.dart';
 import '../ui/details/details_view.dart';
 import '../ui/form/example_form_view.dart';
 import '../ui/home/home_view.dart';
+import '../ui/nonreactive/nonreactive_view.dart';
 import '../ui/stream_view/stream_counter_view.dart';
 
 class Routes {
@@ -21,13 +22,15 @@ class Routes {
   static const String bottomNavExample = '/bottom-nav-example';
   static const String streamCounterView = '/stream-counter-view';
   static const String detailsView = '/details-view';
-  static const String exampleFormView = '/';
+  static const String exampleFormView = '/example-form-view';
+  static const String nonReactiveView = '/';
   static const all = <String>{
     homeView,
     bottomNavExample,
     streamCounterView,
     detailsView,
     exampleFormView,
+    nonReactiveView,
   };
 }
 
@@ -40,6 +43,7 @@ class StackedRouter extends RouterBase {
     RouteDef(Routes.streamCounterView, page: StreamCounterView),
     RouteDef(Routes.detailsView, page: DetailsView),
     RouteDef(Routes.exampleFormView, page: ExampleFormView),
+    RouteDef(Routes.nonReactiveView, page: NonReactiveView),
   ];
   @override
   Map<Type, StackedRouteFactory> get pagesMap => _pagesMap;
@@ -80,6 +84,12 @@ class StackedRouter extends RouterBase {
       );
       return MaterialPageRoute<dynamic>(
         builder: (context) => ExampleFormView(key: args.key),
+        settings: data,
+      );
+    },
+    NonReactiveView: (data) {
+      return MaterialPageRoute<dynamic>(
+        builder: (context) => const NonReactiveView(),
         settings: data,
       );
     },