+++ packages/stacked/lib/src/state_management/reactive_service_mixin.dart
@@ -14,6 +14,8 @@ mixin ReactiveServiceMixin {
         reactiveValue.onChange.listen((event) => notifyListeners());
       } else if (reactiveValue is RxSet) {
         reactiveValue.onChange.listen((event) => notifyListeners());
+      } else if (reactiveValue is ChangeNotifier) {
+        (reactiveValue as ChangeNotifier).notifyListeners();
       }
     }
   }