+++ packages/stacked_generator/lib/src/generators/getit/dependency_config.dart
@@ -5,8 +5,14 @@ class DependencyConfig {
   /// The import to use for the type of the service
   final String import;
 
+  /// The import to use for the abstracted service type
+  final String abstractedImport;
+
   /// The actual name of the class to be registered
   final String className;
+
+  /// The abstracted class name of the class to be registered
+  final String abstractedTypeClassName;
 
   /// The type of the service to register
   final DependencyType type;
@@ -14,11 +20,17 @@ class DependencyConfig {
   /// The static function to use for presolving the service
   final String presolveFunction;
 
+  /// The static function to use for resolving a singleton instance
+  final String resolveFunction;
+
   DependencyConfig({
     this.import,
+    this.abstractedImport,
     this.className,
+    this.abstractedTypeClassName,
     this.type,
     this.presolveFunction,
+    this.resolveFunction,
   });
 
   /// Returns a camel case version o the [className]