+++ packages/stacked_generator/lib/src/generators/getit/stacked_locator_content_generator.dart
@@ -38,19 +38,30 @@ class StackedLocatorContentGenerator extends BaseGenerator {
 
   String _getLocatorRegistrationStringForType(
       DependencyConfig dependencyDefinition) {
+    final hasAbstratedType =
+        dependencyDefinition.abstractedTypeClassName != null;
+    final abstractionType = hasAbstratedType
+        ? '<${dependencyDefinition.abstractedTypeClassName}>'
+        : '';
+
+    final hasResolveFunction = dependencyDefinition.resolveFunction != null;
+    final singletonInstanceToReturn = hasResolveFunction
+        ? '${dependencyDefinition.className}.${dependencyDefinition.resolveFunction}()'
+        : '${dependencyDefinition.className}()';
+
     switch (dependencyDefinition.type) {
       case DependencyType.LazySingleton:
-        return 'locator.registerLazySingleton(() => ${dependencyDefinition.className}());';
+        return 'locator.registerLazySingleton$abstractionType(() => $singletonInstanceToReturn);';
       case DependencyType.PresolvedSingleton:
         return '''
         final ${dependencyDefinition.camelCaseClassName} = await ${dependencyDefinition.className}.${dependencyDefinition.presolveFunction}();
-        locator.registerSingleton(${dependencyDefinition.camelCaseClassName});
+        locator.registerSingleton$abstractionType(${dependencyDefinition.camelCaseClassName});
         ''';
       case DependencyType.Factory:
-        return 'locator.registerFactory(() => ${dependencyDefinition.className}());';
+        return 'locator.registerFactory$abstractionType(() => ${dependencyDefinition.className}());';
       case DependencyType.Singleton:
       default:
-        return 'locator.registerSingleton(${dependencyDefinition.className}());';
+        return 'locator.registerSingleton$abstractionType($singletonInstanceToReturn);';
     }
   }
 
@@ -59,6 +70,7 @@ class StackedLocatorContentGenerator extends BaseGenerator {
     final imports = <String>{"package:stacked/stacked.dart"};
 
     imports.addAll(services.map((e) => e.import));
+    imports.addAll(services.map((e) => e.abstractedImport));
 
     var validImports = imports.where((import) => import != null).toSet();
     var dartImports =