+++ packages/stacked_generator/lib/src/generators/getit/stacked_locator_generator.dart
@@ -5,7 +5,6 @@ import 'package:source_gen/source_gen.dart';
 
 import 'package:stacked/stacked_annotations.dart';
 import 'package:stacked_generator/import_resolver.dart';
-import 'package:stacked_generator/src/core/logger.dart';
 import 'package:stacked_generator/src/generators/enums/dependency_type.dart';
 import 'package:stacked_generator/src/generators/getit/dependency_config.dart';
 import 'package:stacked_generator/src/generators/getit/stacked_locator_content_generator.dart';
@@ -15,8 +14,6 @@ import 'package:stacked_generator/utils.dart';
 const stackedRouteChecker = TypeChecker.fromRuntime(StackedRoute);
 
 class StackedLocatorGenerator extends GeneratorForAnnotation<StackedApp> {
-  final log = getLogger('LocatorGenerator');
-
   @override
   dynamic generateForAnnotatedElement(
     Element element,
@@ -50,42 +47,61 @@ class StackedLocatorGenerator extends GeneratorForAnnotation<StackedApp> {
     var dependencyReader = ConstantReader(dependencyConfig);
     // Get the type of the service that we want to register
     final dependencyClassType = dependencyReader.read('classType').typeValue;
+    final dependencyAbstractedClassType =
+        dependencyReader.peek('asType')?.typeValue;
 
     throwIf(
       dependencyClassType == null,
-      '🛑 No depedency class Type defined for ${dependencyConfig.toString()}. Please make sure that any of the services provided to the services list in the StackedApp annotation has a service provided. Please see the documentation for stacked_generator if you don\'t know what this means.',
+      '🛑 No dependency class Type defined for ${dependencyConfig.toString()}. Please make sure that any of the services provided to the services list in the StackedApp annotation has a service provided. Please see the documentation for stacked_generator if you don\'t know what this means.',
     );
 
     final classElement = dependencyClassType.element as ClassElement;
+    final abstractedClassElement =
+        dependencyAbstractedClassType?.element as ClassElement;
 
     throwIf(
       classElement == null,
-      '🛑 ${dependencyClassType.getDisplayString()} is not a class element. All services should be classes. We don\'t register individual values for global access through the locator. Make sure the value provided as your service type is a class.',
+      '🛑 ${toDisplayString(dependencyClassType)} is not a class element. All services should be classes. We don\'t register individual values for global access through the locator. Make sure the value provided as your service type is a class.',
     );
 
     // Get the import of the class type that's defined for the service
     final import = importResolver.resolve(classElement);
+    final abstractedImport = importResolver.resolve(abstractedClassElement);
 
-    final className = dependencyClassType.getDisplayString();
+    final className = toDisplayString(dependencyClassType);
+
+    final abstractedTypeClassName = dependencyAbstractedClassType != null
+        ? toDisplayString(dependencyAbstractedClassType)
+        : null;
 
     // NOTE: This can be used for actual dependency inject. We do service location instead.
     final constructor = classElement.unnamedConstructor;
 
     final serviceType = _getDependencyType(dependencyReader);
+
     String presolveFunction;
     if (serviceType == DependencyType.PresolvedSingleton) {
-      final presolveUsing = dependencyReader.read('presolveUsing');
-      final presolveObject = presolveUsing.objectValue.toFunctionValue();
-      presolveFunction = presolveObject.displayName;
-      log.d(
-          'presolveUsing:$presolveUsing objectValue:${presolveObject.toString()} presolveFunction:$presolveFunction');
+      final presolveUsing = dependencyReader.peek('presolveUsing');
+      final presolveObject = presolveUsing?.objectValue?.toFunctionValue();
+      presolveFunction = presolveObject?.displayName;
+    }
+
+    String resolveFunction;
+    if (serviceType == DependencyType.LazySingleton ||
+        serviceType == DependencyType.Singleton) {
+      final resolveUsing = dependencyReader.peek('resolveUsing');
+      final resolveObject = resolveUsing?.objectValue?.toFunctionValue();
+      resolveFunction = resolveObject?.displayName;
     }
 
     return DependencyConfig(
       className: className,
+      abstractedTypeClassName: abstractedTypeClassName,
       import: import,
+      abstractedImport: abstractedImport,
       type: serviceType,
       presolveFunction: presolveFunction,
+      resolveFunction: resolveFunction,
     );
   }