+++ packages/stacked_firebase_auth/CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.1.3
+
+- Replaced `flutter_facebook_login` with `flutter_facebook_auth`
+
+## 0.1.2
+
+Updates dependencies
+
+- `firebase_core`: ^1.0.1
+- `firebase_auth`: ^1.0.1
+
+## 0.1.1
+
+- Adds `nonce` to remove the probability of a replay attack when using `AppleSignIn`
+
 ## 0.1.0+1
 
 - Readme formatting update