+++ packages/stacked_generator/CHANGELOG.md
@@ -1,5 +1,29 @@
 # ChangeLog
 
+## 0.2.6
+
+- Adds the `resolveUsing` code generation for `Singleton` and `LazySingleton` registrations
+
+## 0.2.5
+
+- Removes dependency on logger
+
+## 0.2.4
+
+- Bumps `analyzer` and `build_runner` versions
+
+## 0.2.3
+
+- Fixes issue #240
+
+## 0.2.2
+
+- Adds support to register a service `asType` when generating the getIt registrations
+
+## 0.2.1
+
+- fixes the incompaitility with using an analyzer version that requires the withNullability argument on .getDisplayString()
+
 ## 0.2.0
 
 - Adds form generation functionality