+++ packages/stacked_hooks/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.4
+
+- `flutter_hooks`: ^0.15.0 -> ^0.16.0
+- `provider`: ^4.3.1 -> ^5.0.0
+
 ## 0.1.3+2
 
 - Update flutter_hooks 0.15.0